\name{Cin_ex}
\docType{data}
\alias{Cin_ex}
\title{
Carbon Input Data
}
\description{
Fictional carbon input for a model with five soil pools. Columns stand for pools and rows for simulation time steps.
}
\usage{Cin_ex}
\format{A matrix of 5 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

