\name{meas_data_ex}
\docType{data}
\alias{meas_data_ex}
\title{
Measured Data Input
}
\description{
Fictional input data. Contains a matrix of three columns. Column 1: Measuremnt time. Column 2: SOC [t/ha]. Column 3: SON [t/ha].
}
\usage{meas_data_ex}
\format{A matrix containing 3 columns}
\seealso{
  \code{\link{sorcering}}.
}

