\name{Nin_ex}
\docType{data}
\alias{Nin_ex}
\title{
Nitrogen Input Data
}
\description{
Fictional nitrogen input for a model with five soil pools. Columns stand for pools and rows for simulation time steps.
}
\usage{Nin_ex}
\format{A matrix of 5 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

