\name{RothC_Cin_ex}
\docType{data}
\alias{RothC_Cin_ex}
\title{
Carbon Input Data for RothC
}
\description{
Fictional carbon input for the RothC SOC model. Columns stand for pools and rows for simulation time steps.
}
\usage{RothC_Cin_ex}
\format{A matrix of 5 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

