\name{RothC_env_in_ex}
\docType{data}
\alias{RothC_env_in_ex}
\title{
Environmental Input Data for RothC
}
\description{
Fictional environmental input for RothC. Rows stand for simulation time steps. Column 1: atmospheric temperature [Celsius degrees]. Column 2: precipitation [mm].
Column 3: evapotranspiration [mm]. Column 4: zeros or ones, the latter indicate time steps with growing crops.
}
\usage{RothC_env_in_ex}
\format{A matrix of 4 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

