\name{RothC_Nin_ex}
\docType{data}
\alias{RothC_Nin_ex}
\title{
Nitrogen Input Data for RothC
}
\description{
Fictional nitrogen input for the RothC SOC model. Columns stand for pools and rows for simulation time steps.
}
\usage{RothC_Nin_ex}
\format{A matrix of 5 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

