\name{RothC_Nin_ex_sl_spin}
\docType{data}
\alias{RothC_Nin_ex_sl_spin}
\title{
Nitrogen Input Data for RothC using multiple sites with spinup run
}
\description{
Fictional nitrogen input for the RothC SOC model. The input data constists of a list of 3 matrices. Matrix columns stand for pools and matrix rows for simulation time steps.
}
\usage{RothC_Nin_ex_sl_spin}
\format{A list of 3 matrices with 5 columns and 12 rows each}
\seealso{
  \code{\link{sorcering}}.
}

