\name{RothC_site_ex}
\docType{data}
\alias{RothC_site_ex}
\title{
Environmental Input Data for RothC
}
\description{
Fictional site information for RothC. Contains information to calculate xi. Vector of length 4: sample depth [mm], clay content [%], black sand status (0
or 1, 0 if unknown or if black sand method is not desired) and CN ratio (0 if
unknown, but then either C0 and N0 must be defined or calcC0 = TRUE and
calcN0 = TRUE).
}
\usage{RothC_site_ex}
\format{A vector containing 4 numeric entries}
\seealso{
  \code{\link{sorcering}}.
}

