\name{RothC_xi_ex}
\docType{data}
\alias{RothC_xi_ex}
\title{
Environmental Factors Data for RothC
}
\description{
Fictional environmental factors for the RothC SOC model. Columns stand for pools and rows for simulation time steps.
}
\usage{RothC_xi_ex}
\format{A matrix of 5 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

