\name{xi_ex}
\docType{data}
\alias{xi_ex}
\title{
Environmental Factors Data
}
\description{
Fictional environmental factors for a model with five soil pools. Columns stand for pools and rows for simulation time steps.
}
\usage{xi_ex}
\format{A matrix of 5 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

