\name{Yasso_Cin_ex_wood_u_sl}
\docType{data}
\alias{Yasso_Cin_ex_wood_u_sl}
\title{
Carbon Input Data for Yasso using Multiple Sites
}
\description{
Fictional carbon input for the Yasso SOC model (Yasso15 or Yasso20). 
The input data constists of a list of 3 lists, each representing a site and containing 2 sublists. Each sublist represents a wood input layer and contains 4 matrices. Each matrix represents an uncertainty repetition and has 5 columns and 60 rows. Columns stand for pools and rows for simulation time steps.
}
\usage{Yasso_Cin_ex_wood_u_sl}
\format{A list of 3 lists with 2 sublists each, each sublist containing 4 matrices with 5 columns and 60 rows}
\seealso{
  \code{\link{sorcering}}.
}

