\name{Yasso_N0_ex_sl}
\docType{data}
\alias{Yasso_N0_ex_sl}
\title{
Initial Soil Organic Nitrogen Data for Yasso using Multiple Sites
}
\description{
Fictional initial soil organic nitrogen for the Yasso SOC model (any Yasso version). The initial data constists of a list of 3 vectors, each containing five numeric entries, one for each model pool.
}
\usage{Yasso_N0_ex_sl}
\format{A list of 3 vectors, each containing five numeric entries}
\seealso{
  \code{\link{sorcering}}.
}

